Move random.go file to the pkg layer
