+++ pkg/util/random.go
@@ -15,7 +15,7 @@
 // specific language governing permissions and limitations
 // under the License.
 
-package cmdutil
+package util
 
 import (
 	"math/rand"
